% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaLE.R
\name{getDeltaLE}
\alias{getDeltaLE}
\title{Get Delta in Life Exp}
\usage{
getDeltaLE(gap_country)
}
\arguments{
\item{gap_country}{character, a country. See gapminder::gapminder$title}
}
\value{
numeric vector
}
\description{
Uses the gapminder data set to...
}
